Refactor syntax highlight `Pager`
